MCAPP Version 0.2a
Description:

An updated version of the app. Developed ontop of Sean's improvements
and additions, the code for the classes have been tidied up as well as
the comments being sorted. There are some pieces that still need to be
done over the weekend, a final version will be uploaded by Sunday.

Change Log:
- Removed reliance and instances of Note.
- Corrected Comments
- Added Methods to Basic Classes

Known Issues:
- Method Conventions still need renaming
- Implementation of Import/Export as well as a central Data file still
needs to be implemented.