MCAPP 0.8.1
Changes:

- Added a Global Boolean for the eventual log in check (Might become
irrelevant.
- Added various XML classes for the eventual integration of a pop-up

Notes:

Right. I believe I have caught up with the work load. I have added a
global boolean (not in use) for the moment where we check if the person
has logged into the app for the first time.

This might all be irrlevant however depending on the next bit. I have
also integrated xml files and included code into the Display class, with
a new 'Transparent' class to create a pop up box. If it works ... we
could just use a button as this demo shows or we could just trigger the
window by using the boolean.

I need one of you guys to basically integrate this for me. So much has
been done to the UI already it is best not for me to start messing
around with it. If either of you can implement this before Tomorrow, I
am sure Dave would be fairly satisfied with our progress!